FR-10557 - Create proxy single request handler instead of handler per request
